reorganized tests
