+++ src/test/resources/it/ipfs-test/pom.xml
@@ -39,7 +39,7 @@
             <plugin>
                 <groupId>org.dbpedia.databus</groupId>
                 <artifactId>databus-maven-plugin</artifactId>
-                <version>2.0-SNAPSHOT</version>
+                <version>${pluginVersion}</version>
                 <configuration>
                     <deployRepoURL>http://localhost:8081/</deployRepoURL>
                     <packageDirectory>src/test/resources/sampleProj/test-set-deploy</packageDirectory>
