+++ src/test/resources/it/ipfs-test/test-cert-bundle-with-password.p12
