+++ src/test/resources/it/props-config/pom.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>testGroup</groupId>
+    <artifactId>testArtifact</artifactId>
+    <version>1.0</version>
+
+    <properties>
+            <databus.deployRepoURL>http://localhost:8081/</databus.deployRepoURL>
+            <databus.packageDirectory>src/test/resources/sampleProj/test-set-deploy</databus.packageDirectory>
+            <databus.downloadUrlPath>http://pa</databus.downloadUrlPath>
+            <databus.publisher>https://neradis.github.io/test-id.ttl#this</databus.publisher>
+            <databus.pkcs12File>test-cert-bundle.p12</databus.pkcs12File>
+            <databus.license>http://purl.oclc.org/NET/rdflicense/cc-by-sa3.0</databus.license>
+            <databus.documentation><![CDATA[No real version, we also use this as a dataset to test the framework]]></databus.documentation>
+            <deploy.ipfsSettings.containerName>${cn}</deploy.ipfsSettings.containerName>
+            <deploy.ipfsSettings.isInDocker>true</deploy.ipfsSettings.isInDocker>
+            <deploy.ipfsSettings.projectRootDockerPath>/export/sample/</deploy.ipfsSettings.projectRootDockerPath>
+    </properties>
+
+    <parent>
+        <groupId>org.dbpedia.databus</groupId>
+        <artifactId>super-pom</artifactId>
+        <version>1.3-SNAPSHOT</version>
+    </parent>
+    <packaging>pom</packaging>
+
+    <repositories>
+        <repository>
+            <id>archiva.internal</id>
+            <name>Internal Release Repository</name>
+            <url>http://databus.dbpedia.org:8081/repository/internal</url>
+        </repository>
+        <repository>
+            <id>archiva.snapshots</id>
+            <name>Internal Snapshot Repository</name>
+            <url>http://databus.dbpedia.org:8081/repository/snapshots</url>
+            <snapshots>
+                <updatePolicy>always</updatePolicy>
+            </snapshots>
+        </repository>
+    </repositories>
+
+
+    <modules>
+        <module>test-set</module>
+    </modules>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.dbpedia.databus</groupId>
+                <artifactId>databus-maven-plugin</artifactId>
+                <version>${pluginVersion}</version>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
