+++ src/test/resources/it/props-config/test-set/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>testGroup</groupId>
+        <artifactId>testArtifact</artifactId>
+        <version>1.0</version>
+    </parent>
+
+    <groupId>testGroup</groupId>
+    <artifactId>test-set</artifactId>
+    <version>2020.06.05</version>
+
+    <packaging>jar</packaging>
+
+</project>
