+++ src/test/scala/org/dbpedia/databus/test/CommonMavenPluginTest.scala
@@ -8,17 +8,17 @@
  * it under the terms of the GNU Affero General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.dbpedia.databus
+package org.dbpedia.databus.test
 
 import java.io.File
 import java.nio.file.{Path, Paths}
@@ -33,12 +33,13 @@ import org.apache.maven.plugin.testing.AbstractMojoTestCase
 import org.apache.maven.plugin.testing.stubs.MavenProjectStub
 import org.apache.maven.shared.utils.ReaderFactory
 import org.codehaus.plexus.configuration.DefaultPlexusConfiguration
-import org.dbpedia.databus.LogInterceptor.{Debug, Err, Info, LogMessage, Warn}
+import org.dbpedia.databus.DatabusMojo
+import org.dbpedia.databus.test.LogInterceptor._
 import org.mockserver.integration.ClientAndServer
-import org.scalatest.mockito.MockitoSugar
 import org.mockserver.integration.ClientAndServer._
 import org.mockserver.model.HttpRequest._
 import org.mockserver.model.HttpResponse._
+import org.scalatest.mockito.MockitoSugar
 
 
 abstract class CommonMavenPluginTest extends AbstractMojoTestCase with MockitoSugar {
