+++ src/test/scala/org/dbpedia/databus/test/it/CommonMavenPluginIT.scala
@@ -18,18 +18,26 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.dbpedia.databus
+package org.dbpedia.databus.test.it
 
 import java.nio.file.Path
 
 import org.apache.maven.it.Verifier
+import org.dbpedia.databus.DatabusPluginVersion
+import org.dbpedia.databus.test.MockHttpServerOps
 import org.scalatest.FunSuiteLike
 
+/**
+ * Created by .
+ */
 trait CommonMavenPluginIT extends MockHttpServerOps with FunSuiteLike {
 
   def projectPath: Path
 
-  def initVerifier =
-    new Verifier(projectPath.toAbsolutePath.toString)
+  def initVerifier = {
+    val ver = new Verifier(projectPath.toAbsolutePath.toString)
+    ver.setSystemProperty("pluginVersion", DatabusPluginVersion.toString)
+    ver
+  }
 
 }
